Slight improvement over GoogleDNS

+++ googleDNS.sh
@@ -1,24 +1,26 @@
 #!/bin/bash
 
 # Add Google DNS nameservers as main ones to access otherwise unavailable stuff
-# If `setup` script is run, it requires sudo (as it should)
 
-read -p "Do you want to add Google DNS or remove it? (on/off) " reply
+if [ "$(id -u)" != "0" ]; then
+   echo "This script must be run as root" 1>&2
+   exit 1
+fi
 
-if [[ $reply = on ]]
+if [[ ! -f /etc/resolv.conf.head ]]
 then
+    echo "Appending Google nameservers to /etc/resolv.conf.head"
 	# IPv4
 	echo "nameserver 8.8.8.8" >> /etc/resolv.conf.head
 	echo -e "nameserver 8.8.4.4\n" >> /etc/resolv.conf.head
 	# IPv6
 	echo "nameserver 2001:4860:4860::8888" >> /etc/resolv.conf.head
 	echo "nameserver 2001:4860:4860::8844" >> /etc/resolv.conf.head
-
 	sv reload wpa_supplicant # not sure if reloading dhcpcd is necessary
-elif [[ $reply = off ]]
-then
-	rm /etc/resolv.conf.head
-	sv reload wpa_supplicant && sv reload dhcpcd
 else
-	echo "Wrong input"
+	echo "The output of /etc/resolv.conf.head is"
+	cat /etc/resolv.conf.head
+	read -n 1 -p "/etc/resolv.conf.head is going to be removed; if you want to proceed, type (y)es " reply && echo ""
+ 	[ $reply = "Y" -o $reply = "y" ] && rm /etc/resolv.conf.head && sv reload wpa_supplicant && sv reload dhcpcd && exit 0
+	echo "The reply is not (y)es, or something else went wrong" && exit 1
 fi
